requirements.txt
flask
gunicorn==20.0.4

google-api-python-client
google-auth-httplib2
google-auth-oauthlib
werkzeug
uuid
oauth2client
apiclient
requests
